backend/requirements.txt
fastapi
uvicorn
gtts
pydub
speechrecognition
deep-translator
python-dotenv
openai
